requirements.txt
Django==2.1.2
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
psycopg2==2.7.5
pytz==2018.5
